backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "backend"
version = "0.1.1"
description = ""
authors = [{ name = "Anton Bushin", email = "i@abushin.ru" }]
requires-python = "~=3.11"
readme = "README.md"
dependencies = [
    "aiogram>=3.15.0,<4",
    "aiogram-media-group>=0.5.1,<0.6",
    "alembic>=1.14.0,<2",
    "asyncpg>=0.30.0,<0.31",
    "colorama>=0.4.6,<0.5",
    "fastapi[all]>=0.115.12,<0.116",
    "jinja2>=3.1.4,<4",
    "opencv-python-headless>=4.11.0.86,<5",
    "orjson>=3.10.12,<4",
    "pydantic==2.9.2",
    "pydantic-settings>=2.6.1,<3",
    "pymupdf>=1.25.3,<2",
    "pypdf[image]>=5.3.0,<6",
    "pypdfium2>=4.30.1,<5",
    "python-json-logger>=3.2.1,<4",
    "pyzbar>=0.1.9,<0.2",
    "sqlalchemy>=2.0.36,<3",
    "tomlkit>=0.13.2,<0.14",
    "uvicorn[standard]>=0.34.0,<0.35",
    "greenlet>=3.2.2,<4",
    "pyjwt>=2.10.1",
    "passlib[bcrypt]>=1.7.4",
]

[dependency-groups]
dev = [
    "aioresponses>=0.7.8,<0.8",
    "aresponses>=2.1.6,<2.2.0",
    "asgi-lifespan>=2.1.0,<3",
    "faker>=37.0.0,<38",
    "pypng>=0.20220715.0,<0.20220716",
    "pyqrcode>=1.2.1,<2",
    "pytest>=8.3.4,<9",
    "pytest-asyncio>=0.25.3,<0.26",
    "pytest-cov>=6.0.0,<7",
    "pytest-factoryboy>=2.7.0,<3",
    "pytest-lazy-fixtures>=1.1.2,<2",
    "pytest-unordered>=0.6.1,<0.7",
    "pytest-unused-fixtures>=0.2.0,<0.3",
    "pytz>=2023.3,<2024.0",
    "ruff>=0.8.3,<0.9",
    "toml-sort>=0.24.2,<0.25",
    "pytest-freezer>=0.4.9,<0.5",
    "pytest-mock>=3.14.0,<4",
    "uv>=0.7.8,<0.8",
]

[tool.coverage.report]
fail_under = 80
show_missing = true
skip_empty = true

[tool.uv]
package = false

[tool.coverage.run]
branch = true
omit = [
    "src/arepl/*",
]
source = [
    "src",
]

[tool.pytest.ini_options]
addopts = "-s -v -q"
asyncio_mode = "auto"
log_cli = true
log_cli_format = "%(message)s"
log_cli_level = "CRITICAL"
log_file = "logs/pytest.log"
log_file_date_format = "%Y-%m-%d %H:%M:%S"
log_file_format = "[%(asctime)s][%(levelname)s] [%(name)s]: %(message)s (%(pathname)s:%(lineno)s)"
log_file_level = "DEBUG"
testpaths = [
    "tests",
]

[tool.ruff]
exclude = [
    "alembic",
    "migrations",
]
indent-width = 4
line-length = 99
src = ["src", "tests"]
target-version = "py311"

[tool.ruff.format]
docstring-code-format = false
docstring-code-line-length = "dynamic"
indent-style = "space"
line-ending = "auto"
quote-style = "double"
skip-magic-trailing-comma = false

[tool.ruff.lint]
fixable = ["ALL"]
ignore = [
    "D100",  # Missing docstring in public module
    "D101",  # Missing docstring in public class
    "D102",  # Missing docstring in public method
    "D103",  # Missing docstring in public function
    "D104",  # Missing docstring in public package
    "D107",  # Missing docstring in `__init__`
    "D203",  # 1 blank line required before class docstring
    "D212",  # Multi-line docstring summary should start at the first line
]
select = [
    # "D", # pydocstyle
    "B",
    "C",
    "E",
    "F",
    "Q",
    "T",
    "W",
]
unfixable = []

[tool.ruff.lint.per-file-ignores]
"src/arepl/*" = ["T201"]
"tests/mocks/ofd_responses.py" = ["E501"]

[tool.tomlsort]
all = true
ignore_case = true
in_place = true
# no_comments = true
# no_header_comments = true
# no_footer_comments = true
# no_inline_comments = true
# no_block_comments = true
# no_sort_tables = true
# sort_first = ["key1", "key2"]
# sort_table_keys = true
# sort_inline_tables = true
# sort_inline_arrays = true
spaces_before_inline_comment = 2
spaces_indent_inline_array = 4
trailing_comma_inline_array = true
